Move initialization code to main again, fix crash because of Settings not initialized early enough
